+++ src/main/java/mdteam/ait/client/renderers/exteriors/ExteriorRenderer.java
@@ -106,13 +106,17 @@ public class ExteriorRenderer<T extends ExteriorBlockEntity> implements BlockEnt
         // -------------------------------------------------------------------------------------------------------------------
 
         matrices.multiply(RotationAxis.POSITIVE_X.rotationDegrees(180f));
-
-        if (entity.getTardis().get().isSiegeMode()) {
-            if (siege == null) siege = new SiegeModeModel(SiegeModeModel.getTexturedModelData().createModel());
-            siege.renderWithAnimations(entity, this.siege.getPart(), matrices, vertexConsumers.getBuffer(AITRenderLayers.getEntityTranslucentCull(SiegeModeModel.TEXTURE)), maxLight, overlay, 1, 1, 1, 1);
-            matrices.pop();
-            return;
+        try {
+            if (entity.getTardis().get().isSiegeMode()) {
+                if (siege == null) siege = new SiegeModeModel(SiegeModeModel.getTexturedModelData().createModel());
+                siege.renderWithAnimations(entity, this.siege.getPart(), matrices, vertexConsumers.getBuffer(AITRenderLayers.getEntityTranslucentCull(SiegeModeModel.TEXTURE)), maxLight, overlay, 1, 1, 1, 1);
+                matrices.pop();
+                return;
+            }
+        } catch (Exception e) {
+            AITMod.LOGGER.error("Failed to render siege mode", e);
         }
+
 
         String name = entity.getTardis().get().getHandlers().getStats().getName();
         if (name.equalsIgnoreCase("grumm") || name.equalsIgnoreCase("dinnerbone")) {