+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 - Fixed a crash in server tardis manager
 - Add a music section for sounds
 - Added drifting song by radio (I'll update with the new version when he adds that)
+- Added a music disc for drifting
+- Starts playing drifting a random amount of time after the tardis power turns off
+- Fixed a crash with exterior renderer with sodium
+- Made it so that drifting starts playing a random amount of time after power is shut off in a tardis
+- Replaced DeltaTimeManager with the one from neptunelib